carpenterbackend/bookings/serializers.py
from rest_framework import serializers
from .models import Booking, ContactMessage

class BookingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Booking
        fields = "__all__"

class ContactMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactMessage
        fields = "__all__"



