carpenterbackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
multidict==6.7.0
packaging==25.0
propcache==0.4.1
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
twilio==9.8.6
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
yarl==1.22.0
